requirements.txt
pyyaml==6.0
ua-parser==0.18.0
user-agents==2.2.0
Django==3.2.23
django-shortuuidfield==0.1.3
gunicorn==21.2.0
pip==24.0
cryptocompare==0.7.6
whitenoise==6.2.0
psycopg2==2.9.7
Pillow==9.4.0
pytz==2022.7
schedule==1.2.1
resend==0.6.0
django-cors-headers==4.3.0
django-pipeline== 2.1.0
django-crontab== 0.7.1